+++ app/app/schemas/category.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 class Category(BaseModel):
     title: str
     borrow_limit: int
-    borrow_price_per_day: int
+    borrow_price_per_day: float
 
 
 class CategoryCreate(Category):
@@ -14,4 +14,4 @@ class CategoryCreate(Category):
 class CategoryUpdate(Category):
     title: str | None
     borrow_limit: int | None
-    borrow_price_per_day: int | None
+    borrow_price_per_day: float | None
